# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Erdos 服务端 API 契约
  version: 0.1.0-draft
  description: >
    客户端 / 引擎 / 服务端三方的共同语言。契约冻结项：错误码（x-error-codes）、
    表结构、签名格式——冻结后变更必须走契约评审（研发手册「质量标准」红线）。
    本文件为草案（x-contract-status: draft）：已 100% 覆盖实际实现（56 URL，2026-10-02
    契约一致性自检），待 ARCH 评审冻结为 v1。自检入口：server/scripts/contract_audit.py。
  x-contract-status: draft
servers:
  - url: https://api.erdos.example.com
...
paths:
...
  /v1/audit/events:
    post:
      tags: [audit]
      operationId: createAuditEvent
      summary: 写入审计事件（admin / operator）
      description: >
        幂等写接口：必须携带 Idempotency-Key 请求头，重复键只入账一次（40901）。
        角色限 admin|operator：actor_type/actor_id 全部来自请求体，若允许任意登录用户写入，
        等于任何人可冒名落审计记录，"关键操作全审计"的可信度不成立。
        服务端自身的登录/购买/权益变更/许可签发/离线对账留痕由内部埋点写入，不走本接口。
      security:
        - BearerAuth: []
      parameters:
        - name: Idempotency-Key
          in: header
          required: true
          description: 幂等键（调用方生成，长度 1~64）
          schema: { type: string, minLength: 1, maxLength: 64 }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AuditEventCreate" }
      responses:
        "200":
          description: 写入成功
          content:
            application/json:
              schema: { $ref: "#/components/schemas/EnvelopeAuditEventView" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "400":
          description: 参数校验失败（BAD_REQUEST）
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Envelope" }
        "409":
          description: 幂等键重复（CONFLICT）
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Envelope" }
        "429":
          description: 触发限流（RATE_LIMITED）
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Envelope" }
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: JWT 双令牌（SP2-2 接入；SP2-1 网关仅做鉴权透传）
  responses:
    Unauthorized:
      description: 未认证（UNAUTHENTICATED）
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Envelope" }
    Forbidden:
      description: 角色不足（PERMISSION_DENIED）
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Envelope" }
...
  schemas:
...
    Envelope:
      type: object
      description: 统一响应信封。code=0 表示成功；其它取值与 x-error-codes 一一对应。
      required: [code, message, request_id, timestamp]
      properties:
        code:
          type: integer
          description: 业务码，0 表示成功，非 0 见 x-error-codes
        message:
          type: string
          description: 面向使用者的提示文案
        detail:
          type: string
          description: 可选的错误上下文（仅失败时携带）
          nullable: true
        data:
          description: 业务数据（成功时携带）
          nullable: true
        request_id:
          type: string
          description: 请求追踪 ID，与响应头 X-Request-Id 一致
        timestamp:
          type: string
          format: date-time
          description: 服务端 UTC 时间
...
    EnvelopeAuditEventView:
      allOf:
        - $ref: "#/components/schemas/Envelope"
        - type: object
          properties:
            data: { $ref: "#/components/schemas/AuditEventView" }
...
    AuditEventCreate:
      type: object
      additionalProperties: false
      required: [actor_type, actor_id, action, resource_type]
      properties:
        actor_type: { type: string, minLength: 1, maxLength: 32, description: 操作者类型（user/device/admin） }
        actor_id: { type: string, minLength: 1, maxLength: 64 }
        action: { type: string, minLength: 1, maxLength: 64, example: account.login }
        resource_type: { type: string, minLength: 1, maxLength: 32 }
        resource_id: { type: string, maxLength: 64, nullable: true }
        detail:
          type: object
          nullable: true
          description: 结构化详情；禁止携带凭据/密钥
        client_ip: { type: string, maxLength: 45, nullable: true }
    AuditEventView:
      type: object
      required: [id, actor_type, actor_id, action, resource_type, created_at]
      properties:
        id: { type: integer }
        actor_type: { type: string }
        actor_id: { type: string }
        action: { type: string }
        resource_type: { type: string }
        resource_id: { type: string, nullable: true }
        detail: { type: object, nullable: true }
        client_ip: { type: string, nullable: true }
        created_at: { type: string, format: date-time }

# 错误码契约（冻结项）：code = HTTP 状态 * 100 + 序号；新增只能追加，禁止复用与修改语义。
...